Add slack:deploy:rollback task
This addresses issue #19. Here, we are adding a `slack:deploy:rollback`
task that hooks into the Capistrano `deploy:finishing_rollback` task.
This task includes all the same variables available to similar tasks
like starting, finished, and failed.

Additionally, in order to differentiate between a deployment finishing
and a rollback finishing, the `slack:deploy:finished` task was moved to
the Capistrano `deploy:finishing` task.