send slack message before deploy:starting
This adds another message fired before deploy:starting. It keeps the
existing deploy:updating message. The reasoning is that some pre
deployment tasks can take awhile (ie. gracefully shutting down
services) and this can introduce a delay between the start of the actual
deploy and when the first message is posted to Slack.

See #93 for impetus.